requirements.txt
feedparser
requests
openai
loguru
python-dateutil
notebooklm-py